+++ cyclops-ui/src/components/pages/EditModule/custom.css
@@ -1,0 +1,28 @@
+.ant-col.ant-form-item-label {
+  padding: 0 !important;
+}
+
+label.ant-form-item-required:before {
+  bottom: 1em !important;
+  position: relative !important;
+}
+
+code {
+  color: #000080;
+}
+
+.ant-collapse-content-box {
+  padding-top: 12px;
+  border: 1px solid #d3d3d3;
+  border-top: none;
+  border-left: 3px solid #faca93;
+  border-bottom-left-radius: 7px;
+  border-bottom-right-radius: 7px;
+  background-color: #ffffff;
+}
+
+.custom {
+  .ant-collapse-content-box {
+    padding: 0;
+  }
+}
