+++ cyclops-ui/src/components/pages/ModuleDetails.tsx
@@ -46,6 +46,7 @@ import {
 import { gvkString } from "../../utils/k8s/gvk";
 import { mapResponseError } from "../../utils/api/errors";
 import Secret from "../k8s-resources/Secret";
+import defaultTemplate from "../../static/img/default-template-icon.png";
 const languages = [
   "javascript",
   "java",
@@ -577,7 +578,15 @@ const ModuleDetails = () => {
                   src={module.iconURL}
                 />
               ) : (
-                <></>
+                <img
+                  alt=""
+                  style={{
+                    maxHeight: "1.5em",
+                    maxWidth: "1.5em",
+                    marginRight: "12px",
+                  }}
+                  src={defaultTemplate}
+                />
               )}
               {module.name}
             </Title>
