+++ cyclops-ui/src/components/pages/NewModule/NewModule.tsx
@@ -1,4 +1,4 @@
-import React, { useCallback, useEffect, useState } from "react";
+import React, { useEffect, useState } from "react";
 import {
   Alert,
   Button,
@@ -8,27 +8,33 @@ import {
   Form,
   Input,
   InputNumber,
-  notification,
   Row,
   Select,
-  Spin,
   Switch,
-  Tooltip,
   Typography,
+  Tooltip,
+  Modal,
+  Spin,
+  notification,
 } from "antd";
 import axios from "axios";
 import { useNavigate } from "react-router";
 import {
-  InfoCircleOutlined,
-  LockFilled,
   MinusCircleOutlined,
   PlusOutlined,
-  UnlockFilled,
+  InfoCircleOutlined,
 } from "@ant-design/icons";
+import {
+  fileExtension,
+  findMaps,
+  flattenObjectKeys,
+} from "../../../utils/form";
+import "./custom.css";
+import defaultTemplate from "../../../static/img/default-template-icon.png";
+
+import YAML from "yaml";
 
 import AceEditor from "react-ace";
-
-import { useParams } from "react-router-dom";
 
 import "ace-builds/src-noconflict/theme-github";
 
@@ -37,76 +43,77 @@ import "ace-builds/src-noconflict/mode-toml";
 import "ace-builds/src-noconflict/mode-javascript";
 import "ace-builds/src-noconflict/mode-typescript";
 import "ace-builds/src-noconflict/snippets/yaml";
-import { fileExtension, findMaps, flattenObjectKeys } from "../../utils/form";
-import "./custom.css";
-import { numberInputValidators } from "../../utils/validators/number";
-import { stringInputValidators } from "../../utils/validators/string";
-import { templateRef } from "../../utils/templateRef";
+import { numberInputValidators } from "../../../utils/validators/number";
+import { stringInputValidators } from "../../../utils/validators/string";
+import { Option } from "antd/es/mentions";
 import {
   FeedbackError,
   FormValidationErrors,
-} from "../errors/FormValidationErrors";
-import { mapResponseError } from "../../utils/api/errors";
-import {
-  getTemplate,
-  getTemplateInitialValues,
-  Template,
-} from "../../utils/api/template";
+} from "../../errors/FormValidationErrors";
+import { mapResponseError } from "../../../utils/api/errors";
 
 const { Title } = Typography;
 const layout = {
   wrapperCol: { span: 16 },
 };
 
-interface module {
+interface templateStoreOption {
   name: string;
-  values: any;
-  template: templateRef;
+  ref: {
+    repo: string;
+    path: string;
+    version: string;
+  };
 }
 
-const EditModule = () => {
-  const [module, setModule] = useState<module>({
+const NewModule = () => {
+  const [loading, setLoading] = useState(false);
+  const [config, setConfig] = useState({
     name: "",
-    values: {},
-    template: {
-      repo: "",
-      path: "",
-      version: "",
-      resolvedVersion: "",
-    },
-  });
-
-  const [previousValues, setPreviousValues] = useState();
-
-  const [form] = Form.useForm();
-  const [editTemplateForm] = Form.useForm();
-
-  const [initialValuesRaw, setInitialValuesRaw] = useState({});
-
-  const [values, setValues] = useState({});
-  const [isChanged, setIsChanged] = useState(false);
-  const [isTemplateChanged, setIsTemplateChanged] = useState(false);
-  const [config, setConfig] = useState<Template>({
-    name: "",
-    resolvedVersion: "",
+    version: "",
+    manifest: "",
     root: {
       properties: [],
       required: [],
     },
+    dependencies: [],
   });
 
-  const [templateRef, setTemplateRef] = useState<templateRef>({
+  const [template, setTemplate] = useState({
     repo: "",
     path: "",
     version: "",
-    resolvedVersion: "",
   });
-  const [templateRefLock, setTemplateRefLock] = useState(true);
+
+  const [initialValues, setInitialValues] = useState({});
+  const [initialValuesRaw, setInitialValuesRaw] = useState({});
 
   const [error, setError] = useState({
     message: "",
     description: "",
   });
+
+  const [loadingTemplate, setLoadingTemplate] = useState(false);
+  const [loadingTemplateInitialValues, setLoadingTemplateInitialValues] =
+    useState(false);
+
+  const [activeCollapses, setActiveCollapses] = useState(new Map());
+  const updateActiveCollapses = (k: string[], v: any) => {
+    let kk = new Array(k);
+    setActiveCollapses(new Map(activeCollapses.set(kk.join(""), v)));
+  };
+
+  var initLoadedFrom: string[];
+  initLoadedFrom = [];
+  const [newFile, setNewFile] = useState("");
+  const [loadedFrom, setLoadedFrom] = useState(initLoadedFrom);
+  const [loadedValues, setLoadedValues] = useState("");
+  const [loadingValuesFile, setLoadingValuesFile] = useState(false);
+  const [loadingValuesModal, setLoadingValuesModal] = useState(false);
+
+  const [templateStore, setTemplateStore] = useState<templateStoreOption[]>([]);
+
+  const history = useNavigate();
 
   const [notificationApi, contextHolder] = notification.useNotification();
   const openNotification = (errors: FeedbackError[]) => {
@@ -118,20 +125,17 @@ const EditModule = () => {
     });
   };
 
-  const [loadValues, setLoadValues] = useState(false);
-  const [loadTemplate, setLoadTemplate] = useState(false);
+  const [form] = Form.useForm();
 
-  const [activeCollapses, setActiveCollapses] = useState(new Map());
-  const updateActiveCollapses = (k: string[], v: any) => {
-    let kk = new Array(k);
-    setActiveCollapses(new Map(activeCollapses.set(kk.join(""), v)));
-  };
+  useEffect(() => {
+    loadTemplateStore();
+  }, []);
 
-  const history = useNavigate();
+  useEffect(() => {
+    form.validateFields(flattenObjectKeys(initialValues));
+  }, [initialValues, form]);
 
-  let { moduleName } = useParams();
-
-  const mapsToArray = useCallback((fields: any[], values: any): any => {
+  const mapsToArray = (fields: any[], values: any): any => {
     let out: any = {};
     fields.forEach((field) => {
       let valuesList: any[] = [];
@@ -211,191 +215,174 @@ const EditModule = () => {
     });
 
     return out;
-  }, []);
-
-  useEffect(() => {
-    const fetchModuleData = async () => {
-      axios
-        .get("/api/modules/" + moduleName)
-        .then(async (res) => {
-          editTemplateForm.setFieldsValue({
-            repo: res.data.template.repo,
-            path: res.data.template.path,
-            version: res.data.template.version,
-          });
-          setLoadValues(true);
-
-          setTemplateRef({
-            repo: res.data.template.repo,
-            path: res.data.template.path,
-            version: res.data.template.version,
-            resolvedVersion: res.data.template.resolvedVersion,
-          });
-
-          let result = await getTemplate(
-            res.data.template.repo,
-            res.data.template.path,
-            res.data.template.resolvedVersion,
-          );
-
-          if (result.success) {
-            setConfig(result.template);
-            let values = mapsToArray(
-              result.template.root.properties,
-              res.data.values,
-            );
-
-            setModule({
-              name: res.data.name,
-              values: values,
-              template: res.data.template,
-            });
-            form.setFieldsValue(values);
-            setValues(values);
-            setPreviousValues(res.data.values);
-          } else {
-            console.log(result);
-            setError(result.error);
-          }
-
-          setLoadTemplate(true);
-        })
-        .catch((error) => {
-          setError(mapResponseError(error));
-          setLoadTemplate(true);
-          setLoadValues(true);
-        });
-    };
-    fetchModuleData();
-  }, [editTemplateForm, form, moduleName, mapsToArray]);
-
-  useEffect(() => {
-    form.validateFields(flattenObjectKeys(values));
-  }, [values, form]);
-
-  const handleValuesChange = (changedValues: any, allValues: any) => {
-    if (JSON.stringify(allValues) === JSON.stringify(values)) {
-      setIsChanged(false);
-    } else {
-      setIsChanged(true);
-    }
-
-    setValues(allValues);
   };
-
-  const handleTemplateRefChange = (
-    repo: string,
-    path: string,
-    version: string,
-    resolvedVersion: string,
-  ) => {
-    let newTemplate: templateRef = {
-      repo: repo,
-      path: path,
-      version: version,
-      resolvedVersion: resolvedVersion,
-    };
-
-    if (JSON.stringify(module.template) === JSON.stringify(newTemplate)) {
-      setIsTemplateChanged(false);
-    } else {
-      setIsTemplateChanged(true);
-    }
-  };
-
-  async function handleSubmitTemplateEdit(templateEditValues: any) {
-    setLoadTemplate(false);
-
-    let currentValues = form.getFieldsValue();
-
-    let templateResult = await getTemplate(
-      templateEditValues.repo,
-      templateEditValues.path,
-      templateEditValues.version,
-    );
-
-    if (!templateResult.success) {
-      setConfig({
-        name: "",
-        resolvedVersion: "",
-        root: { properties: [], required: [] },
-      });
-      setLoadTemplate(true);
-      setError(templateResult.error);
-      return;
-    }
-
-    let initialValuesResult = await getTemplateInitialValues(
-      templateEditValues.repo,
-      templateEditValues.path,
-      templateEditValues.version,
-    );
-
-    if (!initialValuesResult.success) {
-      setConfig({
-        name: "",
-        resolvedVersion: "",
-        root: { properties: [], required: [] },
-      });
-      setLoadTemplate(true);
-      setError(templateResult.error);
-      return;
-    }
-
-    // both requests successful
-    setTemplateRef({
-      repo: templateEditValues.repo,
-      path: templateEditValues.path,
-      version: templateEditValues.version,
-      resolvedVersion: templateResult.template.resolvedVersion,
-    });
-    handleTemplateRefChange(
-      templateEditValues.repo,
-      templateEditValues.path,
-      templateEditValues.version,
-      templateResult.template.resolvedVersion,
-    );
-
-    setConfig(templateResult.template);
-
-    let mergedValues = findMaps(
-      templateResult.template.root.properties,
-      currentValues,
-      initialValuesResult.initialValues,
-    );
-
-    let mergedValuesMapped = mapsToArray(
-      templateResult.template.root.properties,
-      mergedValues,
-    );
-
-    setValues(mergedValuesMapped);
-
-    setInitialValuesRaw(initialValuesResult.initialValues);
-    form.setFieldsValue(mergedValuesMapped);
-
-    setLoadTemplate(true);
-  }
 
   const handleSubmit = (values: any) => {
-    if (isTemplateChanged) {
-      values = findMaps(config.root.properties, values, initialValuesRaw);
-    } else {
-      values = findMaps(config.root.properties, values, previousValues);
-    }
+    const moduleName = values["cyclops_module_name"];
+
+    values = findMaps(config.root.properties, values, initialValuesRaw);
 
     axios
-      .post(`/api/modules/update`, {
+      .post(`/api/modules/new`, {
+        name: moduleName,
         values: values,
-        name: module.name,
-        template: templateRef,
+        template: {
+          repo: template.repo,
+          path: template.path,
+          version: template.version,
+        },
       })
       .then((res) => {
         window.location.href = "/modules/" + moduleName;
       })
       .catch((error) => {
+        setLoading(false);
         setError(mapResponseError(error));
       });
+  };
+
+  const loadTemplate = async (repo: string, path: string, commit: string) => {
+    setConfig({
+      name: "",
+      version: "",
+      manifest: "",
+      root: {
+        properties: [],
+        required: [],
+      },
+      dependencies: [],
+    });
+    form.setFieldsValue({});
+    setInitialValuesRaw({});
+    setInitialValues({});
+
+    setActiveCollapses(new Map());
+    setLoadingTemplate(true);
+    setLoadingTemplateInitialValues(true);
+
+    setError({
+      message: "",
+      description: "",
+    });
+
+    if (repo.trim() === "") {
+      setError({
+        message: "Invalid repository name",
+        description: "Repository name must not be empty",
+      });
+      setLoadingTemplate(false);
+      setLoadingTemplateInitialValues(false);
+      return;
+    }
+
+    let tmpConfig: any = {};
+
+    await axios
+      .get(
+        `/api/templates?repo=` + repo + `&path=` + path + `&commit=` + commit,
+      )
+      .then((templatesRes) => {
+        setConfig(templatesRes.data);
+        tmpConfig = templatesRes.data;
+
+        setError({
+          message: "",
+          description: "",
+        });
+        setLoadingTemplate(false);
+      })
+      .catch(function (error) {
+        setLoadingTemplate(false);
+        setError(mapResponseError(error));
+      });
+
+    axios
+      .get(
+        `/api/templates/initial?repo=` +
+          repo +
+          `&path=` +
+          path +
+          `&commit=` +
+          commit,
+      )
+      .then((res) => {
+        let initialValuesMapped = mapsToArray(
+          tmpConfig.root.properties,
+          res.data,
+        );
+
+        setInitialValuesRaw(res.data);
+        setInitialValues(initialValuesMapped);
+        form.setFieldsValue(initialValuesMapped);
+
+        setError({
+          message: "",
+          description: "",
+        });
+
+        setLoadingTemplateInitialValues(false);
+      })
+      .catch(function (error) {
+        setLoadingTemplateInitialValues(false);
+        setError(mapResponseError(error));
+      });
+
+    setActiveCollapses(new Map());
+  };
+
+  const loadTemplateStore = async () => {
+    await axios
+      .get(`/api/templates/store`)
+      .then((res) => {
+        setTemplateStore(res.data);
+      })
+      .catch(function (error) {
+        setLoadingTemplate(false);
+        setError(mapResponseError(error));
+      });
+  };
+
+  const findTemplateStoreSelected = (name: string) => {
+    for (let ts of templateStore) {
+      if (ts.name === name) {
+        return ts;
+      }
+    }
+
+    return null;
+  };
+
+  const onTemplateStoreSelected = (v: string) => {
+    const ts = findTemplateStoreSelected(v);
+    if (ts === null) {
+      return;
+    }
+
+    setTemplate({
+      repo: ts.ref.repo,
+      path: ts.ref.path,
+      version: ts.ref.version,
+    });
+
+    loadTemplate(ts.ref.repo, ts.ref.path, ts.ref.version);
+  };
+
+  const onLoadFromFile = () => {
+    setLoadingValuesFile(true);
+    setLoadedValues("");
+
+    if (newFile.trim() === "") {
+      setError({
+        message: "Invalid values file",
+        description: "Values file can't be empty",
+      });
+      return;
+    }
+
+    setLoadingValuesModal(true);
+
+    loadValues(newFile);
+    setLoadingValuesFile(false);
   };
 
   const getCollapseColor = (fieldName: string) => {
@@ -427,6 +414,10 @@ const EditModule = () => {
         {...arrayField}
         name={formItemName}
         rules={[{ required: isRequired }]}
+        style={{
+          paddingTop: "8px",
+          marginBottom: "12px",
+        }}
         label={
           <div>
             {field.display_name}
@@ -459,6 +450,10 @@ const EditModule = () => {
       <Form.Item
         {...arrayField}
         name={formItemName}
+        style={{
+          paddingTop: "8px",
+          marginBottom: "12px",
+        }}
         label={
           <div>
             {field.display_name}
@@ -494,49 +489,138 @@ const EditModule = () => {
   };
 
   const arrayInnerField = (
+    fieldName: string,
+    uniqueFieldName: string[],
+    formItemName: any,
     field: any,
-    parentFieldID: string[],
-    parent: string,
     level: number,
-    arrayField: any,
-    remove: Function,
+    header: React.JSX.Element,
   ) => {
-    if (!field.items || typeof field.items === "string") {
+    if (!field.items || field.items.type === "string") {
       return (
-        <Row>
-          <Form.Item
-            style={{ paddingBottom: "0px", marginBottom: "0px" }}
-            wrapperCol={24}
-            {...arrayField}
-            initialValue={field.initialValue}
-            name={[arrayField.name]}
-          >
-            <Input />
-          </Form.Item>
-          <MinusCircleOutlined
-            style={{ fontSize: "16px", paddingLeft: "10px" }}
-            onClick={() => remove(arrayField.name)}
-          />
-        </Row>
+        <Form.Item>
+          <Form.List name={formItemName}>
+            {(arrFields, { add, remove }) => (
+              <div
+                style={{
+                  border: "solid 1px #d3d3d3",
+                  borderRadius: "7px",
+                  padding: "12px",
+                  width: "100%",
+                  backgroundColor: "#fafafa",
+                }}
+              >
+                {arrFields.map((arrField) => (
+                  <Col key={arrField.key}>
+                    <Row>
+                      <Form.Item
+                        style={{
+                          paddingBottom: "0px",
+                          marginBottom: "0px",
+                          width: "80%",
+                        }}
+                        wrapperCol={{ span: 242 }}
+                        {...arrField}
+                        initialValue={field.initialValue}
+                        name={[arrField.name]}
+                      >
+                        <Input />
+                      </Form.Item>
+                      <MinusCircleOutlined
+                        style={{ fontSize: "16px", paddingLeft: "10px" }}
+                        onClick={() => remove(arrField.name)}
+                      />
+                    </Row>
+                  </Col>
+                ))}
+
+                <Form.Item>
+                  <Button
+                    type="dashed"
+                    onClick={() => add()}
+                    block
+                    icon={<PlusOutlined />}
+                  >
+                    Add
+                  </Button>
+                </Form.Item>
+              </div>
+            )}
+          </Form.List>
+        </Form.Item>
       );
     }
-    if (field.items === "object") {
+    if (field.items.type === "object") {
       return (
-        <div>
-          {mapFields(
-            field.items.properties,
-            parentFieldID,
-            "",
-            level + 1,
-            2,
-            arrayField,
-            field.items.required,
-          )}
-          <MinusCircleOutlined
-            style={{ fontSize: "16px" }}
-            onClick={() => remove(arrayField.name)}
-          />
-        </div>
+        <Collapse
+          size={"small"}
+          onChange={function (value: string | string[]) {
+            if (value.length === 0) {
+              updateActiveCollapses(uniqueFieldName, false);
+            } else {
+              updateActiveCollapses(uniqueFieldName, true);
+            }
+          }}
+        >
+          <Collapse.Panel
+            key={fieldName}
+            header={header}
+            style={{
+              borderRadius: "7px",
+              backgroundColor: getCollapseColor(uniqueFieldName.toString()),
+            }}
+            forceRender={true}
+          >
+            <Form.Item>
+              <Form.List name={formItemName}>
+                {(arrFields, { add, remove }) => (
+                  <>
+                    {arrFields.map((arrField) => (
+                      <Col
+                        key={arrField.key}
+                        style={{ padding: 0, paddingBottom: "12px" }}
+                      >
+                        <div
+                          style={{
+                            border: "solid 1px #d3d3d3",
+                            borderRadius: "7px",
+                            padding: "12px",
+                            width: "100%",
+                            backgroundColor: "#fafafa",
+                          }}
+                        >
+                          {mapFields(
+                            field.items.properties,
+                            [...uniqueFieldName, String(arrField.name)],
+                            "",
+                            level + 1,
+                            2,
+                            arrField,
+                            field.items.required,
+                          )}
+                          <MinusCircleOutlined
+                            style={{ fontSize: "16px" }}
+                            onClick={() => remove(arrField.name)}
+                          />
+                        </div>
+                      </Col>
+                    ))}
+                    <Form.Item>
+                      <Button
+                        type="dashed"
+                        onClick={() => add()}
+                        block
+                        icon={<PlusOutlined />}
+                      >
+                        Add
+                      </Button>
+                    </Form.Item>
+                  </>
+                )}
+              </Form.List>
+            </Form.Item>
+          </Collapse.Panel>
+        </Collapse>
       );
     }
   };
@@ -569,6 +653,11 @@ const EditModule = () => {
     required?: string[],
   ) {
     const formFields: {} | any = [];
+
+    if (!fields) {
+      return <></>;
+    }
+
     fields.forEach((field: any) => {
       let fieldName = field.name;
 
@@ -617,6 +706,10 @@ const EditModule = () => {
             <Form.Item
               {...arrayField}
               name={formItemName}
+              style={{
+                paddingTop: "8px",
+                marginBottom: "12px",
+              }}
               label={
                 <div>
                   {field.display_name}
@@ -641,6 +734,10 @@ const EditModule = () => {
               {...arrayField}
               initialValue={field.initialValue}
               name={formItemName}
+              style={{
+                paddingTop: "8px",
+                marginBottom: "12px",
+              }}
               label={
                 <div>
                   {field.display_name}
@@ -658,8 +755,6 @@ const EditModule = () => {
           );
           return;
         case "boolean":
-          let moduleValues: any = values;
-
           let k = [];
           for (const item of parentFieldID) {
             if (item === "") {
@@ -671,7 +766,7 @@ const EditModule = () => {
           k.push(fieldName);
 
           let checked =
-            getValueFromNestedObject(moduleValues, k) === true
+            getValueFromNestedObject(initialValues, k) === true
               ? "checked"
               : "unchecked";
           formFields.push(
@@ -679,6 +774,11 @@ const EditModule = () => {
               initialValue={field.initialValue}
               name={fieldName}
               id={fieldName}
+              valuePropName={checked}
+              style={{
+                paddingTop: "8px",
+                marginBottom: "12px",
+              }}
               label={
                 <div>
                   {field.display_name}
@@ -687,7 +787,6 @@ const EditModule = () => {
                   </p>
                 </div>
               }
-              valuePropName={checked}
             >
               <Switch />
             </Form.Item>,
@@ -727,18 +826,17 @@ const EditModule = () => {
               span={level === 0 ? 16 : 24}
               offset={level === 0 ? 2 : 0}
               style={{
-                paddingBottom: "15px",
-                paddingLeft: "0px",
-                paddingRight: "0px",
+                paddingTop: "8px",
+                paddingBottom: "8px",
                 marginLeft: "0px",
                 marginRight: "0px",
+                paddingLeft: "0px",
+                paddingRight: "0px",
               }}
             >
               <Collapse
                 size={"small"}
-                style={{
-                  borderColor: "#d3d3d3",
-                }}
+                bordered={false}
                 onChange={function (value: string | string[]) {
                   if (value.length === 0) {
                     updateActiveCollapses(uniqueFieldName, false);
@@ -812,66 +910,22 @@ const EditModule = () => {
               span={level === 0 ? 16 : 24}
               offset={level === 0 ? 2 : 0}
               style={{
-                paddingBottom: "15px",
-                marginLeft: "0px",
-                marginRight: "0px",
+                paddingTop: "8px",
+                paddingBottom: "8px",
                 paddingLeft: "0px",
                 paddingRight: "0px",
+                marginLeft: "0px",
+                marginRight: "0px",
               }}
             >
-              <Collapse
-                size={"small"}
-                onChange={function (value: string | string[]) {
-                  if (value.length === 0) {
-                    updateActiveCollapses(uniqueFieldName, false);
-                  } else {
-                    updateActiveCollapses(uniqueFieldName, true);
-                  }
-                }}
-              >
-                <Collapse.Panel
-                  key={fieldName}
-                  header={header}
-                  style={{
-                    borderRadius: "7px",
-                    backgroundColor: getCollapseColor(
-                      uniqueFieldName.toString(),
-                    ),
-                  }}
-                  forceRender={true}
-                >
-                  <Form.List name={formItemName}>
-                    {(arrFields, { add, remove }) => (
-                      <>
-                        {arrFields.map((arrField) => (
-                          <Col key={arrField.key}>
-                            {arrayInnerField(
-                              field,
-                              [...uniqueFieldName, String(arrField.name)],
-                              "",
-                              level + 1,
-                              arrField,
-                              remove,
-                            )}
-                            <Divider />
-                          </Col>
-                        ))}
-
-                        <Form.Item>
-                          <Button
-                            type="dashed"
-                            onClick={() => add()}
-                            block
-                            icon={<PlusOutlined />}
-                          >
-                            Add
-                          </Button>
-                        </Form.Item>
-                      </>
-                    )}
-                  </Form.List>
-                </Collapse.Panel>
-              </Collapse>
+              {arrayInnerField(
+                fieldName,
+                uniqueFieldName,
+                formItemName,
+                field,
+                level + 1,
+                header,
+              )}
             </Col>,
           );
           return;
@@ -881,6 +935,10 @@ const EditModule = () => {
               wrapperCol={{ span: level === 0 ? 16 : 24 }}
               name={fieldName}
               rules={[{ required: isRequired }]}
+              style={{
+                paddingTop: "8px",
+                marginBottom: "12px",
+              }}
               label={
                 <div>
                   {field.display_name}
@@ -890,7 +948,7 @@ const EditModule = () => {
                 </div>
               }
             >
-              <Form.List name={formItemName}>
+              <Form.List name={formItemName} initialValue={[]}>
                 {(fields, { add, remove }) => (
                   <div
                     style={{
@@ -973,80 +1031,92 @@ const EditModule = () => {
     return formFields;
   }
 
-  const formLoading = () => {
-    if (loadTemplate === false || loadValues === false) {
-      return (
-        <Spin tip="Loading" size="large" style={{ alignContent: "center" }} />
+  function renderFormFields() {
+    if (!loadingTemplate && !loadingTemplateInitialValues) {
+      return mapFields(
+        config.root.properties,
+        [],
+        "",
+        0,
+        0,
+        undefined,
+        config.root.required,
       );
     }
 
+    return <Spin size="large" />;
+  }
+
+  const handleCancel = () => {
+    setLoadingValuesModal(false);
+  };
+
+  const handleImportValues = () => {
+    form.setFieldsValue(YAML.parse(loadedValues));
+    setLoadedValues("");
+    setLoadingValuesModal(false);
+  };
+
+  const renderLoadedFromFiles = () => {
+    if (loadedFrom.length === 0) {
+      return;
+    }
+
+    const files: {} | any = [];
+
+    loadedFrom.forEach((value: string) => {
+      files.push(<p>{value}</p>);
+    });
+
     return (
-      <div>
-        {mapFields(
-          config.root.properties,
-          [],
-          "",
-          0,
-          0,
-          undefined,
-          config.root.required,
-        )}
-        <div style={{ textAlign: "right" }}>
-          <Button
-            type="primary"
-            htmlType="submit"
-            name="Save"
-            disabled={!isChanged && !isTemplateChanged}
-          >
-            Save
-          </Button>{" "}
-          <Button
-            htmlType="button"
-            onClick={() => history("/modules/" + moduleName)}
-          >
-            Back
-          </Button>
-        </div>
-      </div>
+      <Collapse
+        ghost
+        items={[
+          {
+            key: "1",
+            label: "Imported values from",
+            children: files,
+          },
+        ]}
+      />
     );
+  };
+
+  const loadValues = (fileName: string) => {
+    axios
+      .get(fileName)
+      .then((res) => {
+        setLoadedValues(res.data);
+        setError({
+          message: "",
+          description: "",
+        });
+        let tmp = loadedFrom;
+        tmp.push(newFile);
+        setLoadedFrom(tmp);
+      })
+      .catch(function (error) {
+        // setLoadingTemplate(false);
+        // setSuccessLoad(false);
+        setError(mapResponseError(error));
+      });
   };
 
   const onFinishFailed = (errors: any) => {
     let errorMessages: FeedbackError[] = [];
     errors.errorFields.forEach(function (error: any) {
+      let key = error.name.join(".");
+      if (error.name.length === 1 && error.name[0] === "cyclops_module_name") {
+        key = "Module name";
+      }
+
       errorMessages.push({
-        key: error.name.join("."),
+        key: key,
         errors: error.errors,
       });
     });
 
     openNotification(errorMessages);
-  };
-
-  const lockButton = () => {
-    if (templateRefLock) {
-      return (
-        <Button
-          type="primary"
-          icon={<LockFilled />}
-          style={{ marginRight: "10px" }}
-          onClick={function () {
-            setTemplateRefLock(false);
-          }}
-        />
-      );
-    }
-
-    return (
-      <Button
-        type="primary"
-        icon={<UnlockFilled />}
-        style={{ marginRight: "10px" }}
-        onClick={function () {
-          setTemplateRefLock(true);
-        }}
-      />
-    );
   };
 
   return (
@@ -1068,79 +1138,10 @@ const EditModule = () => {
       )}
       {contextHolder}
       <Row gutter={[40, 0]}>
-        <Col span={24}>
-          <Title level={2}>{moduleName}</Title>
-        </Col>
-      </Row>
-      <Row gutter={[40, 0]}>
-        <Col span={24}>
-          <Divider orientation="left" orientationMargin="0">
-            Template
-          </Divider>
-          <Row>
-            <Form
-              form={editTemplateForm}
-              layout="inline"
-              autoComplete={"off"}
-              onFinish={handleSubmitTemplateEdit}
-              onFinishFailed={onFinishFailed}
-              style={{ width: "100%" }}
-            >
-              {lockButton()}
-              <Form.Item
-                name={"repo"}
-                style={{ width: "40%", marginRight: "0" }}
-              >
-                <Input placeholder={"Repository"} disabled={templateRefLock} />
-              </Form.Item>
-              <div
-                style={{
-                  width: "15px",
-                  display: "flex",
-                  alignItems: "center",
-                  justifyContent: "center",
-                }}
-              >
-                /
-              </div>
-              <Form.Item
-                name={"path"}
-                style={{ width: "20%", marginRight: "0" }}
-              >
-                <Input placeholder={"Path"} disabled={templateRefLock} />
-              </Form.Item>
-              <div
-                style={{
-                  width: "15px",
-                  display: "flex",
-                  alignItems: "center",
-                  justifyContent: "center",
-                }}
-              >
-                @
-              </div>
-              <Form.Item
-                name={"version"}
-                style={{ width: "20%", marginRight: "0" }}
-              >
-                <Input
-                  placeholder={"Version"}
-                  addonAfter={templateRef.resolvedVersion.substring(0, 7)}
-                  disabled={templateRefLock}
-                />
-              </Form.Item>
-              <Form.Item style={{ paddingLeft: "10px", width: "5%" }}>
-                <Button
-                  type="primary"
-                  htmlType="submit"
-                  loading={!loadTemplate}
-                  disabled={templateRefLock}
-                >
-                  Load
-                </Button>
-              </Form.Item>
-            </Form>
-          </Row>
+        <Col span={23}>
+          <Title style={{ textAlign: "center" }} level={2}>
+            Define Module
+          </Title>
         </Col>
       </Row>
       <Row gutter={[40, 0]}>
@@ -1152,16 +1153,176 @@ const EditModule = () => {
             autoComplete={"off"}
             onFinish={handleSubmit}
             onFinishFailed={onFinishFailed}
-            onValuesChange={handleValuesChange}
           >
             <Divider orientation="left" orientationMargin="0">
-              Edit Module
+              Module template
             </Divider>
-            {formLoading()}
+            <Row>
+              <Col span={16}>
+                <Select
+                  showSearch={true}
+                  onChange={onTemplateStoreSelected}
+                  style={{ width: "100%" }}
+                  placeholder="Select an option"
+                >
+                  {templateStore.map((option: any, index) => (
+                    <Option key={option.name} value={option.name}>
+                      {option.iconURL !== null &&
+                      option.iconURL !== undefined &&
+                      option.iconURL.length !== 0 ? (
+                        <img
+                          alt=""
+                          style={{
+                            maxHeight: "1.5em",
+                            maxWidth: "1.5em",
+                            marginRight: "8px",
+                          }}
+                          src={option.iconURL}
+                        />
+                      ) : (
+                        <img
+                          alt=""
+                          style={{
+                            maxHeight: "1.5em",
+                            maxWidth: "1.5em",
+                            marginRight: "8px",
+                          }}
+                          src={defaultTemplate}
+                        />
+                      )}
+                      {option.name}
+                    </Option>
+                  ))}
+                </Select>
+              </Col>
+            </Row>
+            <Divider orientation="left" orientationMargin="0">
+              Module name
+            </Divider>
+            <Form.Item
+              name="cyclops_module_name"
+              id="cyclops_module_name"
+              label={
+                <div>
+                  Module name
+                  <p style={{ color: "#8b8e91", marginBottom: "0px" }}>
+                    Enter a unique module name
+                  </p>
+                </div>
+              }
+              rules={[
+                {
+                  required: true,
+                  message: "Module name is required",
+                },
+                {
+                  max: 63,
+                  message:
+                    "Module name must contain no more than 63 characters",
+                },
+                {
+                  pattern: /^[a-z0-9](?:[a-z0-9-]*[a-z0-9])?$/, // only alphanumeric characters and hyphens, cannot start or end with a hyphen and the alpha characters can only be lowercase
+                  message:
+                    "Module name must follow the Kubernetes naming convention",
+                },
+              ]}
+              hasFeedback={true}
+              validateDebounce={1000}
+            >
+              <Input />
+            </Form.Item>
+            <Divider orientation="left" orientationMargin="0">
+              Define Module
+            </Divider>
+            {renderFormFields()}
+            <div style={{ textAlign: "right" }}>
+              <Button
+                onClick={function () {
+                  setLoadingValuesModal(true);
+                }}
+                name="Save"
+              >
+                Load values from file
+              </Button>{" "}
+              <Button
+                type="primary"
+                loading={loading}
+                htmlType="submit"
+                name="Save"
+              >
+                Save
+              </Button>{" "}
+              <Button htmlType="button" onClick={() => history("/")}>
+                Back
+              </Button>
+            </div>
           </Form>
         </Col>
       </Row>
+      <Modal
+        title="Values to import"
+        visible={loadingValuesModal}
+        onCancel={handleCancel}
+        onOk={handleImportValues}
+        width={"50%"}
+      >
+        {error.message.length !== 0 && (
+          <Alert
+            message={error.message}
+            description={error.description}
+            type="error"
+            closable
+            afterClose={() => {
+              setError({
+                message: "",
+                description: "",
+              });
+            }}
+            style={{ marginBottom: "20px" }}
+          />
+        )}
+        {renderLoadedFromFiles()}
+        <Input
+          placeholder={"File reference"}
+          style={{ width: "90%", marginBottom: "10px" }}
+          onChange={(value: any) => {
+            setNewFile(value.target.value);
+          }}
+        />
+        {"  "}
+        <Button
+          type="primary"
+          htmlType="button"
+          style={{ width: "9%" }}
+          onClick={onLoadFromFile}
+          loading={loadingValuesFile}
+        >
+          Load
+        </Button>
+        <AceEditor
+          mode={"yaml"}
+          theme="github"
+          fontSize={12}
+          showPrintMargin={true}
+          showGutter={true}
+          highlightActiveLine={true}
+          onChange={setLoadedValues}
+          setOptions={{
+            enableBasicAutocompletion: true,
+            enableLiveAutocompletion: true,
+            enableSnippets: false,
+            showLineNumbers: true,
+            tabSize: 4,
+            useWorker: false,
+          }}
+          style={{
+            height: "25em",
+            width: "100%",
+          }}
+          value={loadedValues}
+        />
+      </Modal>
     </div>
   );
 };
-export default EditModule;
+export default NewModule;
